refactor specConfig creation
